Split out more classes
